Version 1 is complete
In version 1 user is allowed to open a one column excel file and my
program will figure out how many search results each company has, then
sort by high to low.

In future version I would like to include the ability to pick the
search word instead of only using "prison labor". It would be nice to
have the ability to sort by either "hight to low" or "A to Z". Also I
would like to make it so the user can pick the name for the new file
and save it where ever.